+++ app/src/app/components/Toolbar/BrushSizeSelector.tsx
