+++ app/src/app/components/Toolbar/ResetMapButton.tsx
